Small refactor & fixes on resolve API TRN request pages
