+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/SupportTasks/ApiTrnRequests/Index.cshtml
@@ -37,7 +37,9 @@
                     foreach (var result in Model.Results)
                     {
                         <tr class="govuk-table__row" data-testid="task:@result.SupportTaskReference">
-                            <th scope="row" class="govuk-table__header" data-testid="name">@result.FirstName @result.LastName</th>
+                            <th scope="row" class="govuk-table__header" data-testid="name">
+                                <a href="@LinkGenerator.ApiTrnRequestMatches(result.SupportTaskReference)">@result.FirstName @result.LastName</a>
+                            </th>
                             <td class="govuk-table__cell" data-testid="email">@result.EmailAddress</td>
                             <td class="govuk-table__cell" data-testid="requested-on">@result.CreatedOn.ToString(UiDefaults.DateOnlyDisplayFormat)</td>
                             <td class="govuk-table__cell" data-testid="source"><colored-tag>@result.SourceApplicationName</colored-tag></td>
