+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/SupportTasks/ApiTrnRequests/Resolve/Conventions.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using TeachingRecordSystem.SupportUi.Infrastructure.Filters;
 
-namespace TeachingRecordSystem.SupportUi.Pages.SupportTasks.ApiTrnRequests;
+namespace TeachingRecordSystem.SupportUi.Pages.SupportTasks.ApiTrnRequests.Resolve;
 
 public class Conventions : IConfigureFolderConventions
 {
@@ -11,12 +11,6 @@ public class Conventions : IConfigureFolderConventions
             this.GetFolderPathFromNamespace(),
             model =>
             {
-                // The Index page shows a list rather than a single task
-                if (model.RelativePath.EndsWith("/Index.cshtml"))
-                {
-                    return;
-                }
-
                 model.Filters.Add(new CheckSupportTaskExistsFilterFactory(openOnly: true, supportTaskType: SupportTaskType.ApiTrnRequest));
             });
     }
