+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/SupportTasks/ApiTrnRequests/Resolve/Merge.cshtml
@@ -1,5 +1,5 @@
 @page "/support-tasks/api-trn-requests/{supportTaskReference}/merge/{handler?}"
-@model TeachingRecordSystem.SupportUi.Pages.SupportTasks.ApiTrnRequests.Merge
+@model TeachingRecordSystem.SupportUi.Pages.SupportTasks.ApiTrnRequests.Resolve.Merge
 @{
     ViewBag.Title = "Select the details to merge";
 }
